# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "UPDATE DESCRIPTION FIELD",
        "title": "THUS Blogs Backend",
        "contact": {
            "name": "Robert Hawk",
            "email": "robert@the-hawk.us"
        },
        "version": "0.1.0+0"
    },
    "host": "https://the-hawk.us",
    "basePath": "/apis/blogs/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Healthcheck endpoint. Reports which statuses are currently\nrunning and the current API\\'s version number. If critical\nservices are running, it will return 200. If any of the\ncritical services are down, then the endpoint will return 503.",
                "summary": "Check to assure that the service is running.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.ServiceStatus"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/controllers.ServiceStatus"
                        }
                    }
                }
            }
        },
        "/sites": {
            "get": {
                "description": "Lists all blogs that are managed by this backend. Authentication is required to use this endpoint.",
                "summary": "List all blog sites managed by this backend.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Blog"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new blog site to be managed by this backend. Authentication is required to use this endpoint.",
                "summary": "Create a new blog site.",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Blog"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    }
                }
            }
        },
        "/sites/{domain}": {
            "put": {
                "description": "This endpoint is used to update the name or domain name of a managed blog. Authentication is required.",
                "summary": "Edit data pertaining to a managed blog.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Blog"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ControllerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.ControllerError": {
            "type": "object",
            "additionalProperties": true
        },
        "controllers.ServiceStatus": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.Blog": {
            "type": "object",
            "properties": {
                "domain": {
                    "description": "Domain is the domain name where this blog lives on the internet.",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the title of the blog.",
                    "type": "string"
                }
            }
        }
    }
}